Fixing a cast that was causing in overflow in IndexSegment

+++ ambry-store/src/main/java/com.github.ambry.store/IndexValue.java
@@ -105,11 +105,11 @@ class IndexValue {
         offset = new Offset(logSegmentName, value.getLong());
         flags = value.get();
         int expiresAt = value.getInt();
-        expiresAtMs = expiresAt != Utils.Infinite_Time ? (expiresAt * Time.MsPerSec) : Utils.Infinite_Time;
+        expiresAtMs = expiresAt != Utils.Infinite_Time ? ((long) expiresAt * Time.MsPerSec) : Utils.Infinite_Time;
         originalMessageOffset = value.getLong();
         int operationTimeInSecs = value.getInt();
-        operationTimeInMs =
-            operationTimeInSecs != Utils.Infinite_Time ? (operationTimeInSecs * Time.MsPerSec) : Utils.Infinite_Time;
+        operationTimeInMs = operationTimeInSecs != Utils.Infinite_Time ? ((long) operationTimeInSecs * Time.MsPerSec)
+            : Utils.Infinite_Time;
         serviceId = value.getShort();
         containerId = value.getShort();
         break;
